+++ lib/components/image_banner.dart
@@ -1,48 +1,143 @@
+import 'package:dots_indicator/dots_indicator.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:jellyflix/components/desktop_image_banner.dart';
-import 'package:jellyflix/components/mobile_image_banner.dart';
+import 'package:jellyflix/components/image_banner_inner_landscape.dart';
+import 'package:jellyflix/components/image_banner_inner_portrait.dart';
 import 'package:jellyflix/models/screen_paths.dart';
 import 'package:jellyflix/providers/api_provider.dart';
+import 'package:jellyflix/components/global_state.dart';
 import 'dart:io';
+import 'package:async/async.dart';
 
 import 'package:openapi/openapi.dart';
 
-class ImageBanner extends HookConsumerWidget {
+/*
+
+The approach is going to be like this:
+Create the restartable timer in this layer
+create an outer "InteractionArea", passing the timer
+create a 
+*/
+
+class InteractionArea extends StatelessWidget {
+  final void Function()? timerResetCallback;
+  final Function(bool)? setHoveredCallback;
+  final Widget child;
+
+  const InteractionArea({
+    super.key,
+    required this.timerResetCallback,
+    required this.child,
+    this.setHoveredCallback,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return MouseRegion(
+        hitTestBehavior: HitTestBehavior.translucent,
+        onEnter: (_) {
+          timerResetCallback!();
+          setHoveredCallback!(true);
+        },
+        onHover: (value) {
+          // We intentionally don't set hovered here
+          // Otherwise hovering the region with a touchscreen
+          // would break scrolling until rebuild (no exit event)
+          timerResetCallback!();
+        },
+        onExit: (_) {
+          timerResetCallback!();
+          setHoveredCallback!(false);
+        },
+        child: Listener(
+            behavior: HitTestBehavior.translucent,
+            onPointerMove: (event) {
+              timerResetCallback!();
+            },
+            child: child));
+  }
+}
+
+class ImageBanner extends StatefulHookConsumerWidget {
   final List<BaseItemDto> items;
-  final Duration scrollDuration;
+  final Duration scrollInterval;
   final double? height;
 
   const ImageBanner(
       {super.key,
       required this.items,
       this.height = 500,
-      this.scrollDuration = const Duration(seconds: 5)});
+      this.scrollInterval = const Duration(seconds: 5)});
+  @override
+  ImageBannerState createState() => ImageBannerState();
+}
+
+class ImageBannerState extends ConsumerState<ImageBanner> {
+  final PageController _controller = PageController();
+  int _currentPage = 0;
+  RestartableTimer? _timer;
+  bool hovered = false;
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return LayoutBuilder(builder: (context, constraints) {
-      if (Platform.isAndroid || Platform.isIOS) {
-        return MobileImageBanner(
-          items: items,
-          height: height,
-          scrollDuration: scrollDuration,
-          onPressedPlay: onPressedPlay(ref, context),
-        );
-      } else {
-        return DesktopImageBanner(
-          items: items,
-          height: 400,
-          scrollDuration: scrollDuration,
-          onPressedPlay: onPressedPlay(ref, context),
-        );
+  void initState() {
+    super.initState();
+    _timer = RestartableTimer(widget.scrollInterval, () {
+      if (!ref.read(globalState.mediaPlaybackIsLoading) && !hovered) {
+        if (_currentPage < widget.items.length - 1) {
+          _currentPage++;
+        } else {
+          _currentPage = 0;
+        }
+        _controller.animateToPage(_currentPage,
+            duration: const Duration(milliseconds: 350), curve: Curves.easeIn);
       }
+      _timer?.reset();
+    });
+  }
+
+  @override
+  void dispose() {
+    _timer?.cancel();
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(builder: (context, constraints) {
+      return InteractionArea(
+          timerResetCallback: _timer?.reset,
+          setHoveredCallback: (bool value) {
+            setState(() {
+              hovered = value;
+            });
+          },
+          child: MediaQuery.of(context).orientation == Orientation.portrait
+              ? ImageBannerInnerPortrait(
+                  items: widget.items,
+                  height: widget.height,
+                  onPressedPlay: onPressedPlay(ref, context),
+                  controller: _controller,
+                  currentPage: _currentPage,
+                  setCurrentPageCallback: (int currentPage) =>
+                      {setState(() => _currentPage = currentPage)})
+              : ImageBannerInnerLandscape(
+                  items: widget.items,
+                  height: widget.height,
+                  onPressedPlay: onPressedPlay(ref, context),
+                  controller: _controller,
+                  currentPage: _currentPage,
+                  setCurrentPageCallback: (int currentPage) =>
+                      {setState(() => _currentPage = currentPage)}));
     });
   }
 
   onPressedPlay(WidgetRef ref, BuildContext context) {
     return (item) async {
+      ref
+          .read(globalState.mediaPlaybackIsLoading.notifier)
+          .update((state) => true);
       var itemId = item.id;
       var playbackStartTicks = item.userData?.playbackPositionTicks ?? 0;
       if (item.type == BaseItemKind.series) {
